fix!: remove unnecessary `DescribeRule` interface
